Time: 3 ms (46.47%), Space: 6 MB (42.98%) - LeetHub

+++ 50-powx-n/50-powx-n.cpp
@@ -19,7 +19,7 @@ public:
                 nn=nn/2;
             }
         }
-        if(n<0) ans=1.0/ans;
+        if(n<0) ans=double(1.0)/double(ans);
         return ans;
     }
 };
